move the software controller to its location
